+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.5.0 - Unreleased
+## 0.5.0 - 2026-08-26
 
 ### Agent 宿主可靠性